requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
blurhash-python==1.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-nested-routers==0.93.5
drf-spectacular==0.27.1
email-validator==2.1.0.post1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
inflect==7.0.0
inflection==0.5.1
ipython==8.20.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
